requirements.txt
Flask
numpy
pandas
opencv-python
matplotlib
seaborn
tqdm
scikit-learn
tensorflow-directml-plugin